restructure files to feature based rather than component based
