+++ client/package.json
@@ -8,6 +8,8 @@
     "@mui/icons-material": "^5.0.1",
     "@mui/lab": "^5.0.0-alpha.50",
     "@mui/material": "^5.0.1",
+    "@reduxjs/toolkit": "^2.8.2",
+    "@tanstack/react-query": "^4.40.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
@@ -15,6 +17,7 @@
     "formik": "^2.2.9",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-redux": "^8.1.3",
     "react-router-dom": "^5.3.0",
     "react-scripts": "^5.0.1",
     "recharts": "^2.1.4",
