+++ client/src/features/auth/SignIn/SignIn.js
@@ -13,12 +13,12 @@ import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import { useHistory } from 'react-router';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { signInSchema } from '../../schemas/SignInSchema';
+import {signInSchema} from './SignInSchema';
 import { useFormik } from 'formik';
 import { Link } from 'react-router-dom';
 import Swal from 'sweetalert2';
 
-const interactions = require('../../services/dataInteraction');
+const interactions = require('../../../services/dataInteraction');
 const theme = createTheme();
 
 /**
