+++ client/src/features/dashboard/Coin/CoinGraphInterval.js
