+++ client/src/features/dashboard/Coins/Coins.js
