+++ client/src/features/dashboard/Favourite/FavouriteItem.js
@@ -3,7 +3,7 @@ import Typography from '@mui/material/Typography';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import { Link } from 'react-router-dom';
-import Title from '../Title/Title';
+import Title from '../../Title/Title';
 import PropTypes from 'prop-types';
 
 export default function FavouriteItem({ coins }) {
