+++ client/src/features/dashboard/Portfolio/Portfolio.js
@@ -6,9 +6,9 @@ import TableCell from '@mui/material/TableCell';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import TableBody from '@mui/material/TableBody';
-import Title from '../Title/Title';
+import Title from '../../Title/Title';
 import PortfolioList from './PortfolioList';
-import { getData } from '../../services/dataInteraction';
+import { getData } from '../../../services/dataInteraction';
 
 export default function Portfolio() {
   const [transList, setTransList] = useState([]);
